refactor(orc8r): Segregation of Streamer servicer as external
